Almost Finishing
I Made a lot of changes and i'm almost finishing the challenge. There's some details to make, but i think it's closer to the design than i imagined it could be while i did it by my self. These details that are missing could be soulved by using JS tec, i guess, and i don't know how to use it for now.